+++ tiny-bench/src/output/analysis/sample_data.rs
@@ -1,5 +1,7 @@
 use crate::benching::SamplingData;
-use crate::output::analysis::criterion::SamplingDataSimpleAnalysis;
+use crate::output::analysis::criterion::{
+    calculate_median, calculate_variance, SamplingDataSimpleAnalysis,
+};
 
 pub(crate) fn simple_analyze_sampling_data(
     sampling_data: &SamplingData,
@@ -26,12 +28,18 @@ pub(crate) fn simple_analyze_sampling_data(
         total += sample_average;
         total_elapsed += elapsed_nanos;
     }
+    let median = calculate_median(&mut sample_averages);
     let total_average = total / sampling_data.samples.len() as f64;
+    let variance = calculate_variance(&sample_averages, total_average);
+    let stddev = variance.sqrt();
     SamplingDataSimpleAnalysis {
         elapsed: total_elapsed,
         min,
         max,
         average: total_average,
+        median,
+        variance,
+        stddev,
         per_sample_average: sample_averages,
     }
 }