+++ tiny-bench/src/output/mod.rs
@@ -6,7 +6,7 @@ pub(crate) mod ser;
 use crate::benching::SamplingData;
 #[cfg(feature = "bench")]
 use crate::output::analysis::criterion::{
-    calculate_p_value, calculate_t_value, resample, BenchmarkConfig,
+    calculate_p_value, calculate_t_value, resample, BenchmarkConfig, SamplingDataSimpleAnalysis,
 };
 #[cfg(feature = "bench")]
 use crate::output::analysis::sample_data::simple_analyze_sampling_data;
@@ -73,7 +73,7 @@ impl Output for SimpleStdout {
     fn dump_timing_data(&self, label: &'static str, data: TimingData) {
         print_timer_header(label, &data);
         let mean = data.elapsed as f64 / data.iterations as f64;
-        print_elapsed(data.min_nanos as f64, mean, data.max_nanos as f64);
+        timer_print_elapsed(data.min_nanos as f64, mean, data.max_nanos as f64);
     }
 
     #[cfg(feature = "bench")]
@@ -86,7 +86,7 @@ impl Output for SimpleStdout {
     ) {
         let analysis = simple_analyze_sampling_data(sampling_data);
         print_sample_header(label, total_iters, analysis.elapsed, cfg.num_samples as u64);
-        print_elapsed(analysis.min, analysis.average, analysis.max);
+        print_analysis(&analysis);
     }
 }
 
@@ -99,7 +99,7 @@ impl Output for ComparedStdout {
         let mean = data.elapsed as f64 / data.iterations as f64;
         let maybe_old = disk::try_read_last_results(label);
         print_timer_header(label, &data);
-        print_elapsed(data.min_nanos as f64, mean, data.max_nanos as f64);
+        timer_print_elapsed(data.min_nanos as f64, mean, data.max_nanos as f64);
         match maybe_old {
             Ok(Some(old)) => {
                 let min_change = (data.min_nanos as f64 / old.min_nanos as f64 - 1f64) * 100f64;
@@ -141,7 +141,7 @@ impl Output for ComparedStdout {
     ) {
         let analysis = simple_analyze_sampling_data(sampling_data);
         print_sample_header(label, total_iters, analysis.elapsed, cfg.num_samples as u64);
-        print_elapsed(analysis.min, analysis.average, analysis.max);
+        print_analysis(&analysis);
         match disk::try_read_last_simpling(label) {
             Ok(Some(last)) => {
                 let old_analysis = simple_analyze_sampling_data(&last);
@@ -155,7 +155,7 @@ impl Output for ComparedStdout {
                 let t_distribution = resample(
                     &analysis.per_sample_average,
                     &old_analysis.per_sample_average,
-                    100,
+                    cfg.num_resamples,
                 );
                 let p = calculate_p_value(t, &t_distribution);
                 let mean_change = if mean_change.abs() >= NOISE_THRESHOLD && p <= SIGNIFICANCE_LEVEL
@@ -211,7 +211,26 @@ pub(crate) fn print_sample_header(
     );
 }
 
-pub(crate) fn print_elapsed(min: f64, mean: f64, max: f64) {
+#[cfg(feature = "bench")]
+pub(crate) fn print_analysis(analysis: &SamplingDataSimpleAnalysis) {
+    // Variance has the unit T-squared,
+    println!(
+        "\telapsed\t[{} {} {}]:\t[{} {} {}] (sample data: med = {}, var = {}², stddev = {})",
+        wrap_gray("min"),
+        wrap_high_intensity_white("mean"),
+        wrap_gray("max"),
+        wrap_gray(&fmt_time(analysis.min)),
+        wrap_high_intensity_white(&fmt_time(analysis.average)),
+        wrap_gray(&fmt_time(analysis.max)),
+        fmt_time(analysis.median),
+        fmt_time(analysis.variance),
+        fmt_time(analysis.stddev),
+    );
+}
+
+#[cfg(feature = "timer")]
+pub(crate) fn timer_print_elapsed(min: f64, mean: f64, max: f64) {
+    // Variance has the unit T-squared,
     println!(
         "\telapsed\t[{} {} {}]:\t[{} {} {}]",
         wrap_gray("min"),
@@ -219,7 +238,7 @@ pub(crate) fn print_elapsed(min: f64, mean: f64, max: f64) {
         wrap_gray("max"),
         wrap_gray(&fmt_time(min)),
         wrap_high_intensity_white(&fmt_time(mean)),
-        wrap_gray(&fmt_time(max))
+        wrap_gray(&fmt_time(max)),
     );
 }